interceptor/requirements.txt
mitmproxy==10.4.0
httpx==0.27.0
pystray==0.19.5
Pillow==10.4.0

